+++ router/underlayproviders/afpacketudpip/portfilter_test.go
@@ -23,14 +23,29 @@
 package afpacketudpip
 
 import (
+	"net"
 	"testing"
 
+	"github.com/gopacket/gopacket/afpacket"
 	"github.com/stretchr/testify/assert"
 )
 
 func TestRawSocket(t *testing.T) {
 	// Interface #1 is lo0. Open the next one. We don't really care what it is
 	// for this test.
-	_, err := RawSocket(2, 50000)
+	intf, err := net.InterfaceByIndex(2)
+	assert.NoError(t, err)
+
+	afpHandle, err := afpacket.NewTPacket(
+		afpacket.OptInterface(intf.Name),
+		afpacket.OptFrameSize(4096))
+
+	assert.NoError(t, err)
+
+	filter, err := BpfSockFilter(50000)
+	assert.NoError(t, err)
+
+	// Attach the program to the raw socket.
+	err = afpHandle.SetEBPF(int32(filter))
 	assert.NoError(t, err)
 }
